AV-974: Implement new date parser for tablesorter
The tablesorter could not sort Finnish and most of the Swedish dates.
Implement a new parser so we can sort the dates in their current format.

We can take the date string and convert it to a Date object, which in
turn can utilize the getTime() function to turn the string into a numerical
value. The numerical value is easier and faster to sort, thus making
the sorting more efficient and independent of the language selected in
the browser.

In order to acoomplish this, the string date format needs to be converted
into a format which the Date object accepts. The helper functions in report.js
take care of this. Finnish format gets converted to mm/dd/yyyy and the Swedish
date retains its format(same as English) but gets translated into the English.

Refs. AV-974